permanently import dwm 4.7 sources
Because there is no hope of updating the dwm patchsets to a newer
version without completely rewriting them and it has become a PITA to
update this patch monster for even a simple fix, use source control as
it is meant to be and checkin the source directly.

dmenu is still handled via the patchset because I like somehow like
this weird build system I've built and it still is manageable and easy
to update to current dmenu upstream versions.