add experimental widescreen layout
New layout with master window centered and slave windows to the left
and right.  Optimized towards widescreen displays where it can get
annoying if the columns get too wide.

Remainders are not distributed yet: There might be some unused pixels
to the bottom of the slave stacks (particularly to the bottom right
with an odd number of slave windows).