snap clients to local screen borders on mousemove()
For now only snaps to the borders of the screen you started on.
When moving to another screen, drop the client there, pick it up
again and the borders from the new screen will be snapped to.

Determining the correct screen to snap to does not seem that easy ;-)

global desktop borders.

This change does absolutely nothing when you have only one screen.