switch screens when moving/resizing floating clients
When a floating client is moved or resized, it will become part of the
screen it has the most area on.  This prevents confusion when a
floating cleint from screen 1 is moved over to screen 2 and still
listens to workspace switches from screen 1.

(inspired by the INTERSECT macro from upstream dwm)