fix crash with Eclipse Neon in Change Method Signature dialog
dwm: fatal error: request code=2, error code=8
X Error of failed request:  BadMatch (invalid parameter attributes)
  Major opcode of failed request:  2 (X_ChangeWindowAttributes)
  Serial number of failed request:  8660
  Current serial number in output stream:  8661

Looks like a XChangeWindowAttributes event is received for a window
that is already gone, so just add it to the "don't care" list in
xerror().  Seems to be the same case as with XConfigureWindows which
is already in that list.