maximize fix v1
When a floating window appears, it pushes the old selected client back
to the stack and the client that is on the top of the stack will be
displayed maximized instead of the previously selected client.  This
is very irritating, because after closing the floating client (a
dialog window in most cases) you have to hunt through the stack to get
your previously selected client back.

This currently only works for new clients.  Moving a floating client
from one workspace to another one still changes the maximized client
when the target workspace is maximized.