show status bar only once per physical screen
When multiple workspaces share the same Xinerama screen because they
have been created by the HORIZONTALAUTOSPLIT configuration, they get a
new flag split[workspace] set to True.

When the active client is on a workspace with split[workspace] set to
True, the status bar is not drawn on that workspace to conserve space,
as autosplit workspaces tend to get narrow.  Because at least one
additional workspace exists on the same Xinerama (psysical) screen,
the status line is still shown at least once on that screen.

(because split[] is only a boolean, multiple split workspaces on
 multiple Xinerama screens can't be distinguished, but this should be
 of no consequence because only one of those screens can have the
 active client at a given time)