reset split flag in updatexinerama()
When changing from a big to a small resolution, the workspace split
was correctly reverted (only one workspace one the screen),
but the split flag was still set and the statusbar still drawn in
split-mode as introduced in  commit 63175a70.

This change correctly reverts the statusbar to normal behaviour
when split mode is left.